+++ ncov19_dash/__init__.py
@@ -1,0 +1,5 @@
+from .dash_app import app
+from .flask_server import server
+# from ncov19_dash.app import create_app
+
+# app = create_app()
