+++ ncov19_dash/assets/js/index.js
