+++ ncov19_dash/components/column_stats.py
@@ -1,9 +1,8 @@
 import requests
 import dash_bootstrap_components as dbc
 import dash_html_components as html
-from utils import REVERSE_STATES_MAP
-from utils import config
-from app import cache
+from ncov19_dash.utils import REVERSE_STATES_MAP
+from ncov19_dash.utils import config
 import pandas as pd
