+++ ncov19_dash/components/column_stats_mobile.py
@@ -1,8 +1,8 @@
 import requests
 import dash_bootstrap_components as dbc
 import dash_html_components as html
-from utils import config
-from app import cache
+from ncov19_dash.utils import config
+from ncov19_dash.flask_server import cache
 import pandas as pd
