+++ ncov19_dash/components/confirmed_cases_chart.py
@@ -3,8 +3,8 @@ import json
 import pandas as pd
 import plotly.express as px
 import plotly.graph_objects as go
-from app import cache
-from utils import config
+from ncov19_dash.flask_server import cache
+from ncov19_dash.utils import config
 
 
 @cache.memoize(timeout=3600)
