+++ ncov19_dash/components/daily_stats_mobile.py
@@ -4,9 +4,9 @@ import requests
 from typing import List, Dict
 import dash_bootstrap_components as dbc
 import dash_html_components as html
-from app import cache
-from utils import config
-from utils import STATES_COORD
+from ncov19_dash.flask_server import cache
+from ncov19_dash.utils import config
+from ncov19_dash.utils import STATES_COORD
 
 
 def safe_div(x, y):
