+++ ncov19_dash/components/scatter_mapbox.py
@@ -7,9 +7,9 @@ from dash.dependencies import Input, Output, State
 import plotly.express as px
 import plotly.graph_objects as go
 
-from app import cache
-from utils import STATES_COORD
-from utils import config
+from ncov19_dash.flask_server import cache
+from ncov19_dash.utils import STATES_COORD
+from ncov19_dash.utils import config
 import requests
