+++ ncov19_dash/components/state-population-est2019.csv
