+++ ncov19_dash/components/twitter_feed.py
@@ -6,9 +6,9 @@ from dateutil.parser import parse
 import requests
 import dash_bootstrap_components as dbc
 import dash_html_components as html
-from app import cache
-from utils import config
-from utils import STATES_COORD
+from ncov19_dash.flask_server import cache
+from ncov19_dash.utils import config
+from ncov19_dash.utils import STATES_COORD
 
 
 @cache.memoize(timeout=900)
