+++ ncov19_dash/pages/mobile_footer.py
