+++ ncov19_dash/utils/settings.py
