+++ ncov19_dash/wsgi.py
@@ -1,0 +1,7 @@
+from werkzeug.middleware.dispatcher import DispatcherMiddleware
+from .app import server
+from .app import app
+
+application = DispatcherMiddleware(server, {
+    '/': app.server,
+})
